requirements.txt
Flask
Flask-SQLAlchemy
pyserial
Flask-Caching
filelock
python-crontab
